feat: added view controller and settings views to main settings page
